Change carma user and group IDs to 1001
On Jetson platforms, the nvidia user and group are already assigned 1000,
which conflicts with the IDs used for the carma user and group. The carma
user and group IDs can be more easily changed than the nvidia ones. They
can be set to 1001 on the Jetson, but the carma user and group inside the
container must match in order to pass volumes successfully.

Dockerfile
* change carma user and groups to use ID 1001 instead of 1000