Change carma user and group to 1001 in Docker container image (#87)
* 80 - Remove graphics flag for driver capabilities

The NVIDIA_DRIVER_CAPABILITIES flag defaults to `all` on the Jetson TX2, and
it seems that any flags after `all` result in a build failure. This commit
removes extra flags, which resolves the issue. Since the default flag is `all`,
any other flags are redundant and unnecessary.

Dockerfile
* remove `graphics` flag from NVIDIA_DRIVER_CAPABILITIES

* Change carma user and group IDs to 1001

On Jetson platforms, the nvidia user and group are already assigned 1000,
which conflicts with the IDs used for the carma user and group. The carma
user and group IDs can be more easily changed than the nvidia ones. They
can be set to 1001 on the Jetson, but the carma user and group inside the
container must match in order to pass volumes successfully.

Dockerfile
* change carma user and groups to use ID 1001 instead of 1000