Add `robot_state_publisher` package (#89)
* 80 - Remove graphics flag for driver capabilities

The NVIDIA_DRIVER_CAPABILITIES flag defaults to `all` on the Jetson TX2, and
it seems that any flags after `all` result in a build failure. This commit
removes extra flags, which resolves the issue. Since the default flag is `all`,
any other flags are redundant and unnecessary.

Dockerfile
* remove `graphics` flag from NVIDIA_DRIVER_CAPABILITIES

* 84 - Add missing package dependencies

Add a .repos file with dependent packages and add installation commands
in the Dockerfile

Dockerfile
* add commands to clone the dependent package repos and build and install
  them

carma-base.repos
* add non-CARMA package source repos so that they can be installed from
  source in the container image

* Add missing dependency

Add tf2_geometry_msgs package to dependency list. It is needed to build the
amathutils_lib package, which is a dependency for the lanelet2_extension
package.

Dockerfile
* add tf2_geometry_msgs package as a source dependency

* Alphabetize ROS source dependencies

Alphabetize ROS source dependencies to make lookups a bit easier.
This does not have any improvement other than readability

Dockerfile
* Alphabetize ROS source dependencies

* Change carma user and group IDs to 1001

On Jetson platforms, the nvidia user and group are already assigned 1000,
which conflicts with the IDs used for the carma user and group. The carma
user and group IDs can be more easily changed than the nvidia ones. They
can be set to 1001 on the Jetson, but the carma user and group inside the
container must match in order to pass volumes successfully.

Dockerfile
* change carma user and groups to use ID 1001 instead of 1000

* 88 - Add robot_state_publisher package

Add the robot_state_publisher package, which is used by downstream packages
for publishing static transforms.

Dockerfile
* add liburdfdom-dev and liburdfdom-headers-dev system package
  dependencies
* add robot_state_publisher to the package build list

carma-base.repos
* add robot_state_publisher source repo
* add kdl_parser source repo
* add urdf source repo