Add the required dependencies for Autoware.Auto to carma-base (#101)
This PR adds the dependencies required to build the Autoware.Auto packages which may be used by CARMA Platform.
ROS packages are added with apt instead of the package.xml as that is currently used for the Noetic build. Once the fully ROS2 release occurs they can be moved back to that structure.